# User a high-performance attribute detector