|                    Method |       Mean |    Error |   StdDev | Ratio | RatioSD |
|-------------------------- |-----------:|---------:|---------:|------:|--------:|
|                 IsDefined |   653.8 ns | 13.07 ns | 16.53 ns |  1.00 |    0.00 |
|        GetCustomAttribute | 1,149.6 ns | 22.97 ns | 22.56 ns |  1.76 |    0.06 |
| GetGenericCustomAttribute | 1,216.5 ns | 24.15 ns | 54.51 ns |  1.81 |    0.07 |